moved models.py from cli/ to engine/
